Improve clarity of tutorial and timing of sound for bot clink
The tutorial now has two boxes to indicate what it is referring to to make it more clear. The text instructions have also been changed for clarity. The timing of the bot clink has been changed so it aligns more with the animation hitting the bottom of the board. Lastly, the reset button style was updated to match the Pass and Play screen.