Fix bug of bot making bad moves because board state had not been updated
The timeout to determine the bot's move has been extended so that the bot will be determining its move based on the current board, instead of a previous board. There was a bug where the bot would make illegal moves (placing a chip in a column that is full) or missing a win when it could win on its turn.