Fix the bug of bot putting chip in full column
After further testing, the previous commit did not actually fix the bug. The bug resulted from "null > -1" evaluating to true. If the first move the bot considered was illegal, it would make an illegal move because it would compare -1 (the default best score for a move) and null, and determine that the null move is the best. This bug was hard to find because it only happened when the first move that it considered was illegal  or if all moves it considered before the illegal move evaluated to -1.